+++ R/import_peaks_bedgraph.R
@@ -2,15 +2,15 @@ import_peaks_bedgraph <- function(paths,
                                   id,
                                   query_granges, 
                                   build,
-                                  call_peaks_method,
+                                  method,
                                   cutoff,
                                   peaks_dir,
                                   verbose=TRUE){
     #### Import bedGraph subset #### 
-    which <- if(is.null(call_peaks_method)) query_granges else NULL
+    which <- if(is.null(method)) query_granges else NULL
     peaks_all <- lapply(paths, function(x){
         if((!is.null(query_granges)) &
-           (!is.null(call_peaks_method))){
+           (!is.null(method))){
             ## Import the entire chromosome to accurately compute peaks.
             chroms <- as.character(
                 unique(GenomicRanges::seqnames(query_granges))
@@ -26,7 +26,7 @@ import_peaks_bedgraph <- function(paths,
             gr <- rtracklayer::import.bedGraph(con = x, which = which)
         } 
         #### Exit early ####
-        if(is.null(call_peaks_method)) {
+        if(is.null(method)) {
             messager("Returning bedGraph GRanges without computing peaks.",
                      v=verbose)
             GenomicRanges::mcols(gr)$peaktype <- "bedGraph"
@@ -41,7 +41,7 @@ import_peaks_bedgraph <- function(paths,
                                      con = tmp_lifted)
         #### Call peaks ####
         peaks <- call_peaks(bedgraph_path = tmp_lifted,
-                            call_peaks_method = call_peaks_method,
+                            method = method,
                             cutoff = cutoff,
                             outdir = peaks_dir,
                             outputfile = paste(